src/main/java/com/example/shoppingwebapplication/Services/UserServiceImpl.java
package com.example.shoppingwebapplication.Services;

import com.example.shoppingwebapplication.Entity.User;
import com.example.shoppingwebapplication.Repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class UserServiceImpl implements UserService {

    private final UserRepository userRepository;

    @Autowired
    public UserServiceImpl(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    @Override
    public void saveUser(User user) {
        userRepository.save(user);
    }

    @Override
    public User findByUserName(String userName) {
        return userRepository.findByusername(userName);
    }
}
